+++ CHANGELOG.md
@@ -4,10 +4,13 @@ ACF Wizard Changelog
 0.0.2
 -----
  - Feature: Make auto-disabling optional
+ - Feature: Proceed button size
  - Fix: remove var_dump
  - Fix: Prefill values for radio, true_false
  - Fix: Proceed step seletion resetted on load group admin
  - Fix: Prefill values admin styles
+ - Fix: JS Error in Field amin (this.table)
+ - Fix: First step not selected if not in stepper
 
 0.0.1
 -----